Move EventLoopMonitor to shc.interfaces.system_monitoring and make its status subscribable
