+++ shc/supervisor.py
@@ -166,81 +166,6 @@ async def interface_failure(interface_name: str = "n/a") -> None:
     asyncio.create_task(stop())
 
 
-class EventLoopMonitor(AbstractInterface):
-    """
-    A special SHC interface class for monitoring the health of the asyncio Event Loop.
-
-    This interface does not provide any connectors, but only implements the :meth:`get_status` method for allowing
-    external monitoring systems to monitor the health of this application's event loop.
-
-    For this purpose, when started, it regularly checks the current number of asyncio tasks and the delay of scheduled
-    function calls in the event loop. These values are reported in the metrics dict of the interface status. The
-    interface's service status is determined by comparing these metrics to fixed threshold values.
-
-    There should only be single instance of this class, which can be accessed at
-    :var:`shc.supervisor.LOOP_MONITOR_INTERFACE`.
-    """
-    def __init__(self, lag_warning: float = 0.005, lag_error: float = 0.02, tasks_warning: int = 1000,
-                 tasks_error: int = 10000):
-        super().__init__()
-        self.interval = 5.0
-        self.num_aggr_samples = 60
-        self.samples: Deque[Tuple[float, int]] = collections.deque()
-        self.task: asyncio.Task
-        self.tic = 0.0
-
-        self.lag_warning = lag_warning
-        self.lag_error = lag_error
-        self.tasks_warning = tasks_warning
-        self.tasks_error = tasks_error
-
-    async def start(self) -> None:
-        self.task = asyncio.create_task(self._monitor_loop())
-
-    async def stop(self) -> None:
-        self.task.cancel()
-        await self.task
-
-    async def _monitor_loop(self) -> None:
-        loop = asyncio.get_running_loop()
-        while True:
-            self.tic = loop.time()
-            loop.call_soon(self._measure_delay)
-            try:
-                await asyncio.sleep(self.interval)
-            except asyncio.CancelledError:
-                return
-
-    def _measure_delay(self) -> None:
-        lag = asyncio.get_running_loop().time() - self.tic
-        queue_length = sum(1 for task in asyncio.all_tasks() if not task.done())
-        self.samples.append((lag, queue_length))
-        while len(self.samples) > self.num_aggr_samples:
-            self.samples.popleft()
-
-    async def _get_status(self) -> "InterfaceStatus":
-        lag_max, tasks_max = functools.reduce(lambda a, i: (max(a[0], i[0]), max(a[1], i[1])), self.samples, (0.0, 0))
-        warning = lag_max >= self.lag_warning or tasks_max >= self.tasks_warning
-        error = lag_max >= self.lag_error or tasks_max >= self.tasks_error
-        return InterfaceStatus(
-            ServiceStatus.UNKNOWN if len(self.samples) == 0
-            else ServiceStatus.CRITICAL if error
-            else ServiceStatus.WARNING if warning
-            else ServiceStatus.OK,
-            "",
-            {
-                'lag_max': lag_max,
-                'tasks_max': tasks_max
-            }
-        )
-
-    def __repr__(self) -> str:
-        return "EventLoopMonitor"
-
-
-LOOP_MONITOR_INTERFACE = EventLoopMonitor()
-
-
 async def run():
     _SHC_STOPPED.clear()
     logger.info("Starting up interfaces ...")
